@@ -1,51 +1,89 @@
-import os
 from pathlib import Path
 
 from loguru import logger
+from tqdm import tqdm
 
 
-def mkdir_if_not_exists(dir:Path, main_process: bool = True):
+def mkdir_if_not_exists(dir: Path, main_process: bool = True):
+    dir = Path(dir)
     if not dir.exists() and main_process:
-        logger.info(f"Creating directory {dir}...")
+        logger.info(f"Creating directory {dir} ...")
         dir.mkdir(parents=True, exist_ok=True)
 
 
-def download_file(url, filename, force=False):
-    import urllib.request
+def download_file(url, target_path, chunk_size=8192):
+    import requests
 
-    if force or not os.path.exists(filename):
-        logger.info(f"Downloading {url} to {filename}...")
-        urllib.request.urlretrieve(url, filename)
+    target_path = Path(target_path) if isinstance(target_path, str) else target_path
+    existing_size = target_path.stat().st_size if target_path.exists() else 0
+    headers = {"Range": f"bytes={existing_size}-"}
+
+    with requests.get(url, headers=headers, stream=True) as r:
+        total_size = int(r.headers.get("content-length", 0)) + existing_size
+        with (
+            open(target_path, "ab") as f,
+            tqdm(
+                total=total_size,
+                unit="iB",
+                unit_scale=True,
+                desc=f"Downloading {target_path.name}",
+                leave=True,
+                initial=existing_size,
+            ) as pbar,
+        ):
+            for chunk in r.iter_content(chunk_size):
+                size = f.write(chunk)
+                pbar.update(size)
+
+    pbar.set_description(f"Downloaded {target_path.name} ({total_size} bytes)")
+    print()  # Add a newline after tqdm
 
 
 def unzip_file(zip_file, dest_dir):
     import zipfile
 
     mkdir_if_not_exists(dest_dir)
-    logger.info(f"Unzipping {zip_file} to {dest_dir}...")
+    logger.info(f"Unzipping {zip_file} to {dest_dir} ...")
     with zipfile.ZipFile(zip_file, "r") as zip_ref:
         zip_ref.extractall(dest_dir)
+    logger.info(f"Unzipping completed: {zip_file} extracted to {dest_dir}.")
 
 
 def untar_file(tar_file, dest_dir):
     import tarfile
 
     mkdir_if_not_exists(dest_dir)
-    logger.info(f"Extracting  {tar_file} to {dest_dir}...")
+    logger.info(f"Extracting  {tar_file} to {dest_dir} ...")
     with tarfile.open(tar_file, "r|gz") as tar_ref:
         tar_ref.extractall(dest_dir)
+    logger.info(f"Extracting completed: {tar_file} extracted to {dest_dir}.")
 
 
-def download_zenodo_record(zenodo_id: str, target_dir: str, force_download=False):
-    zenodo_archive = f"https://zenodo.org/api/records/{zenodo_id}/files-archive"
-    target_path = f"{target_dir}/{zenodo_id}.zip"
+def download_zenodo_record(zenodo_id: str, target_dir: str, force_download: bool = False, temp_dir: None | str = None):
+    import shutil
+    import tempfile
 
-    if not force_download and Path(target_path).exists():
-        logger.info(f"{target_path} already exists.")
+    target_zip_path = Path(target_dir) / f"{zenodo_id}.zip"
+    if not force_download and target_zip_path.exists():
+        logger.info(f"{target_zip_path} already exists, skipping download.")
     else:
-        download_file(zenodo_archive, target_path, force=force_download)
+        temp_zip_path = Path(tempfile.gettempdir()) / f"{zenodo_id}.zip"
+        zenodo_archive_url = f"https://zenodo.org/api/records/{zenodo_id}/files-archive"
+        download_file(zenodo_archive_url, temp_zip_path)
+        shutil.move(temp_zip_path, target_zip_path)
+        logger.info(f"Downloading completed: {zenodo_id} saved to {target_zip_path}.")
 
-    unzip_file(f"{target_dir}/{zenodo_id}.zip", target_dir)
+    try:
+        unzipped_flag = Path(target_dir) / f".unzipped"
+        if not unzipped_flag.exists():
+            unzip_file(target_zip_path, target_dir)
+            unzipped_flag.touch()
+        else:
+            logger.info(f"{target_zip_path} already unzipped, skipping unzip.")
+    except Exception as e:
+        logger.error(f"Failed to unzip {target_zip_path} to {target_dir}: {e}.")
+        logger.error(f"Remove {target_zip_path} and retry, or download manually using `download_data.sh`.")
+        raise e
 
 
 def attr_from_py_path(path: str, endswith: str | None = None) -> type:
@@ -58,7 +96,6 @@ def attr_from_py_path(path: str, endswith: str | None = None) -> type:
     except ModuleNotFoundError:
         raise ValueError(f"Module not found: {module_name}")
 
-    print(dir(module))
     attr_list = [m for m in dir(module) if not endswith or m.endswith(endswith)]
     if len(attr_list) != 1:
         raise ValueError(f"Expected 1 class with endswith={endswith}, got {len(attr_list)}")
