@@ -1,7 +1,7 @@
 from xares.task import TaskConfig
 
 
-def fsdkaggle2018_config(**kwargs) -> TaskConfig:
+def fsdkaggle2018_config(encoder) -> TaskConfig:
     data_key = "sound"
     class_label_maps = {
         "Hi-hat": 0,
@@ -46,22 +46,17 @@ def fsdkaggle2018_config(**kwargs) -> TaskConfig:
         "Violin_or_fiddle": 39,
         "Finger_snapping": 40,
     }
-
-    config_params = {
-        "name": "fsdkaggle2018",
-        "batch_size_train": 64,
-        "learning_rate": 1e-3,
-        "train_split": "fsd18_train",
-        "test_split": "fsd18_test",
-        "valid_split": "fsd18_test",
-        "zenodo_id": "TODO",
-        "output_dim": len(class_label_maps),
-        "label_processor": lambda x: class_label_maps[x[data_key]],
-        "epochs": 20,
-    }
-
-    config_params.update(kwargs)
-
-    config = TaskConfig(**config_params)
-
-    return config
+    return TaskConfig(
+        name="fsdkaggle2018",
+        encoder=encoder,
+        batch_size_train=64,
+        learning_rate=1e-3,
+        train_split="fsd18_train",
+        test_split="fsd18_test",
+        valid_split="fsd18_test",
+        zenodo_id="14725117",
+        output_dim=len(class_label_maps),
+        label_processor=lambda x: class_label_maps[x[data_key]],
+        do_knn=False,
+        epochs=3,
+    )
