@@ -1,7 +1,7 @@
 from xares.task import TaskConfig
 
 
-def gtzan_genre_config(**kwargs) -> TaskConfig:
+def gtzan_genre_config(encoder) -> TaskConfig:
     class_label_maps = {
         "blues": 0,
         "classical": 1,
@@ -14,16 +14,14 @@ def gtzan_genre_config(**kwargs) -> TaskConfig:
         "reggae": 8,
         "rock": 9,
     }
-    config_params = {
-        "name": "gtzan_genre",
-        "zenodo_id": "TODO",
-        "k_fold_splits": list(range(0, 10)),
-        "output_dim": len(class_label_maps),
-        "label_processor": lambda x: class_label_maps[x["genre"]],
-    }
-    config_params.update(kwargs)
-
-    config = TaskConfig(**config_params)
+    config = TaskConfig(
+        encoder=encoder,
+        name="gtzan_genre",
+        zenodo_id="14722472",
+        k_fold_splits=list(range(0, 10)),
+        output_dim=len(class_label_maps),
+        label_processor=lambda x: class_label_maps[x["genre"]],
+    )
 
     config.audio_tar_name_of_split = {fold: f"gtzan_fold_{fold}_0000000.tar" for fold in config.k_fold_splits}
     config.encoded_tar_name_of_split = {fold: f"gtzan-wds-encoded-fold-{fold}-*.tar" for fold in config.k_fold_splits}
